src/requirements.txt
django
psycopg[binary]
djangorestframework
redis
django-redis
celery[redis]
drf-spectacular
djangorestframework-simplejwt
requests
django-cors-headers
Pillow
